fix: introduce escape utility and use it to escape html attribute and body of injected data script block
